+++ src/components/NotificationHandler/NotificationDescription/index.ts
@@ -1,0 +1,3 @@
+import NotificationDescription from './NotificationDescription';
+
+export default NotificationDescription;