+++ src/components/NotificationHandler/index.ts
@@ -1,0 +1,3 @@
+import NotificationHandler from './NotificationHandler';
+
+export default NotificationHandler;