+++ src/constants/defaults.ts
@@ -1,0 +1,5 @@
+// Notification Durations
+// Antd Notifications, not Antd Messages
+// -------------------------------------
+export const SHORT_MESSAGE_NOTIFICATION_DURATION = 7;
+export const LONG_MESSAGE_NOTIFICATION_DURATION = 15;