+++ src/routes/SQLScheduler/views/index.ts
@@ -1,0 +1,4 @@
+export { default as SQLSchedulerJobForm } from './SQLSchedulerJobForm';
+export { default as SQLSchedulerJobLogs } from './SQLSchedulerJobLogs';
+export { default as SQLSchedulerJobsTable } from './SQLSchedulerJobsTable';
+export { default as SQLSchedulerManageJob } from './SQLSchedulerManageJob';