+++ test/testUtils/actWithFakeTimers.ts
@@ -1,0 +1,7 @@
+const actWithFakeTimers = (action: () => void) => {
+  jest.useFakeTimers();
+  action();
+  jest.runAllTimers();
+};
+
+export default actWithFakeTimers;